Now getting the next forcast interval correctly
There is a 'special' site that may be included with a parameterId/locationID/StationID of 'NextFcst'. This is a site that was created specifically to hold a special entry to indicate the next time the XML data set will be updated. In that site, there multiple events. the event with a flag of '0' is the one we want. We read the 'value' attribute of that event.